// client/stocklinx app/src/redux/supplier/constant.ts
export enum supplierConst {
  FETCH_SUPPLIERS_REQUEST = "FETCH_SUPPLIERS_REQUEST",
  FETCH_SUPPLIERS_SUCCESS = "FETCH_SUPPLIERS_SUCCESS",
  FETCH_SUPPLIERS_FAILURE = "FETCH_SUPPLIERS_FAILURE",
  FETCH_SUPPLIER_REQUEST = "FETCH_SUPPLIER_REQUEST",
  FETCH_SUPPLIER_SUCCESS = "FETCH_SUPPLIER_SUCCESS",
  FETCH_SUPPLIER_FAILURE = "FETCH_SUPPLIER_FAILURE",
  CREATE_SUPPLIER_REQUEST = "CREATE_SUPPLIER_REQUEST",
  CREATE_SUPPLIER_SUCCESS = "CREATE_SUPPLIER_SUCCESS",
  CREATE_SUPPLIER_FAILURE = "CREATE_SUPPLIER_FAILURE",
  UPDATE_SUPPLIER_REQUEST = "UPDATE_SUPPLIER_REQUEST",
  UPDATE_SUPPLIER_SUCCESS = "UPDATE_SUPPLIER_SUCCESS",
  UPDATE_SUPPLIER_FAILURE = "UPDATE_SUPPLIER_FAILURE",
  REMOVE_SUPPLIER_REQUEST = "REMOVE_SUPPLIER_REQUEST",
  REMOVE_SUPPLIER_SUCCESS = "REMOVE_SUPPLIER_SUCCESS",
  REMOVE_SUPPLIER_FAILURE = "REMOVE_SUPPLIER_FAILURE",
  SET_SUPPLIER = "SET_SUPPLIER",
  CLEAR_SUPPLIER = "CLEAR_SUPPLIER",
  SET_SUPPLIERS = "SET_SUPPLIERS",
  CLEAR_SUPPLIERS = "CLEAR_SUPPLIERS",
}